Split layout example and refactored imported code
